Build libcudf with -Wall(#7105)
I discovered we're not building libcudf the `-Wall` GCC flag. This PR enables `-Wall` for GCC and nvcc, and fixes most of the errors.

~~The only error I haven't fixed yet is `-Werror=uninitialized` on this line [this line](https://github.com/rapidsai/cudf/blob/branch-0.18/cpp/include/cudf/scalar/scalar.hpp#L334), but @codereport is on it.~~ Fixed :heavy_check_mark:

Authors:
  - ptaylor <paul.e.taylor@me.com>
  - Conor Hoekstra <codereport@outlook.com>
  - Paul Taylor <paul.e.taylor@me.com>

Approvers:
  - Conor Hoekstra (@codereport)
  - Keith Kraus (@kkraus14)
  - Mark Harris (@harrism)

URL: https://github.com/rapidsai/cudf/pull/7105

+++ cpp/tests/jit/jit-cache-multiprocess-test.cpp
@@ -82,7 +82,7 @@ TEST_F(JitCacheMultiProcessTest, MultiProcessTest)
   CUDA_TRY(cudaMallocManaged(&input, sizeof(input)));
   CUDA_TRY(cudaMallocManaged(&output, sizeof(output) * num_tests * 2));
 
-  for (size_t i = 0; i < num_tests; i++) {
+  for (int i = 0; i < num_tests; i++) {
     if (cpid > 0)
       usleep(10000);
     else
